pyproject.toml
[tool.poetry]
name = "knownepal bot"
version = "1.0"
description = "A Bot for KnowNepal discord"
authors = ["Nishant Sapkota <snishant306@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
hikari = "^2.0.0.dev111"
hikari-lightbulb = "^2.2.4"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
aerich = "^0.7.1"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}


[tool.aerich]
tortoise_orm = "know_nepal.core.tortoise_config.tortoise_config"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
